ai-service/requirements.txt
fastapi
uvicorn
pandas
numpy
mlxtend
pymongo
python-dotenv
statsmodels
scikit-learn
pydantic
